src/backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
mysqlclient
python-decouple
python-dotenv
GitPython
requests
PyMuPDF>=1.23.0
openai
drf-spectacular
pytest
pytest-django
django-cors-headers
Pillow
